fastapi-postgres-model-tdd/pyproject.toml
[tool.poetry]
name = "fastapi-postgres-model-tdd"
version = "0.1.0"
description = ""
authors = ["Igor Martins <igor.mart.san@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
